+++ apps/tool_django_generator/views.py
@@ -10,9 +10,23 @@ from rest_framework import status
 from django.conf import settings
 import os, json, pprint
 
+from django_ratelimit.decorators import ratelimit
+from django.utils.decorators import method_decorator
+
+from apps.common.models import *
+from helpers.generator import * 
+from helpers.util import get_client_ip 
+
+from apps.tasks.tasks import *
+from celery.result import AsyncResult
+from core.celery import celery_app
+from apps.common.models_generator import * 
+
 # Create your views here.
 
+#@ratelimit(key='user_or_ip', rate='3/m')
 def index(request):
+
     context = {
         'segment'        : 'django_generator',
         'parent'         : 'tools',
@@ -23,21 +37,58 @@ def index(request):
     }
     return render(request, "tools/django-generator.html", context)
 
-
 class StatusView(APIView):
 
+    #@method_decorator(ratelimit(key='user_or_ip', rate='3/m'))
     def post(self, request):
-        # Use pprint to print the incoming JSON data from the frontend
-        pprint.pprint(request.data)
+        
+        result = task_generator.delay( request.data )
 
-        response_data = {
-            "status": "200",
-            "info": "Django Template is generating"
-        }
-        return Response(response_data, status=status.HTTP_200_OK)
+        app = GeneratedApp()
+        app.task_id = result.id
+        app.user_ip = get_client_ip( request )
+
+        if request.user.is_authenticated:
+            app.user = request.user
+            print( ' > User ' + str( request.user ) )
+        else:
+            print( ' > Guest User ')
+        
+        # Save the creation
+        app.save()
+
+        '''
+        count = 0
+        while not AsyncResult( result.id ).ready():
+            count += 1
+            time.sleep(1)
+            if count > 999:
+                # kill task
+                celery_app.control.revoke(result.id, terminate=True)
+                # Update status
+                app.task_state = COMMON.CANCELLED
+                app.save()
+        '''
+
+        task_result = result.get()
+
+        app.task_log = json.dumps( task_result ) 
+        if 'gh_repo' in task_result:
+            app.gh_repo = task_result['gh_repo']
+
+        # Save the update
+        app.task_state  = task_result['task_state']
+        app.task_result = task_result['task_result']  
+        app.save()
+
+        task_result['status'] = task_result['task_state'] + ', ' + task_result['task_result']
+        task_result['info'  ] = task_result['task_info'] + ', result: ' + task_result['task_output'] 
+
+        return Response(task_result, status=status.HTTP_200_OK)
 
 
 class DesignView(APIView):
+
     def get(self, request):
         data_file_path = os.path.join(settings.MEDIA_ROOT, 'generator/django', 'data.json')