+++ core/__init__.py
@@ -1,5 +1,5 @@
 from __future__ import absolute_import, unicode_literals
 
-from core.celery import app as celery_app
+from core.celery import celery_app
 
 __all__ = ('celery_app',)